+++ Sources/SheetMusicLayout/Skyline/LayoutElementShape+Base.swift
@@ -279,7 +279,7 @@ extension LayoutElementShape {
         var barLineHalfHeight: CGFloat?
         switch element {
         case let .clef(_, origin, _),
-             let .keySignature(_, _, origin),
+             let .keySignature(_, _, _, origin),
              let .timeSignature(_, _, origin),
              let .measureRepeat(_, origin),
              let .multiMeasureRest(_, origin):